Refactor reload spinner into component
