+++ src/components/Refresh.vue
@@ -1,8 +1,12 @@
 ﻿<template>
 <div id="refresher">
+  <!-- Spinner on page reload -->
+  <b-spinner id="load" label="Loading..."></b-spinner>
+  <!-- Reload button -->
   <b-button class="refresh" id="rb" @click="refreshGraph" variant="info">&#8635;</b-button>
+  <!-- Reload button spinner -->
   <b-button class="refresh" id="rs" variant="info" disabled>
-    <b-spinner small></b-spinner>
+    <b-spinner small style="margin-bottom: 2px"></b-spinner>
     <span class="sr-only">Loading...</span>
   </b-button>
 </div>
@@ -12,7 +16,7 @@
 export default {
 name: "Refresh",
   methods: {
-    displayRefreshLoad (b) { // Used when pushing the Refresh button
+    displayRefreshSpinner (b) { // Used when clicking the Refresh button
       const spin = document.getElementById("rs")
       const button = document.getElementById("rb")
       if (b) {
@@ -21,6 +25,19 @@ name: "Refresh",
       } else {
         spin.style.display = "none"
         button.style.display = "block"
+      }
+    },
+    displayReloadSpinner (b) { // Used when reloading the page (F5)
+      const spinner = document.getElementById("load")
+      const refresh = document.getElementById("rb")
+      if (b) {
+        spinner.style.display = "block"
+        refresh.style.display = "none"
+        return false // Tells App whether the Statusbar should be shown
+      } else {
+        spinner.style.display = "none"
+        refresh.style.display = "block"
+        return true
       }
     },
     refreshGraph () {
@@ -39,5 +56,13 @@ name: "Refresh",
 
 #rs {
   display: none;
+}
+
+#load {
+  position: fixed;
+  left: 48%;
+  bottom: 48%;
+  width: 3rem;
+  height: 3rem;
 }
 </style>
